login, FTP y Recordatorio de contraseña 27/01

+++ src/albergueperronclient/ui/controller/FTPController.java
@@ -5,6 +5,7 @@
  */
 package albergueperronclient.ui.controller;
 
+import albergueperronclient.exceptions.FTPException;
 import albergueperronclient.logic.IFTPImplementation;
 import albergueperronclient.modelObjects.MyFile;
 import static albergueperronclient.ui.controller.GenericController.LOGGER;
@@ -12,6 +13,7 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Supplier;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -24,9 +26,12 @@ import javafx.event.EventHandler;
 import javafx.event.EventType;
 import javafx.fxml.FXML;
 import javafx.scene.Group;
+import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn.CellDataFeatures;
 import javafx.scene.control.TextField;
 import javafx.scene.control.TreeItem;
@@ -35,25 +40,25 @@ import javafx.scene.control.TreeTableView;
 import javafx.scene.control.TreeView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.control.cell.TreeItemPropertyValueFactory;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
 import static javafx.scene.input.KeyCode.T;
 import javafx.scene.input.MouseEvent;
 
 import javafx.scene.text.Text;
 import javafx.stage.FileChooser;
+import javafx.stage.Stage;
 import javafx.stage.WindowEvent;
 import org.apache.commons.net.ftp.FTPClient;
 import org.apache.commons.net.ftp.FTPFile;
 
 
 /**
- * Controller UI class for Login view in users' management application. It contains 
- * event handlers and initialization code for the view defined in Login.fxml file.
- * @author javi
+
+ * @author Nerea Jimenez
  */
 public class FTPController extends GenericController{
-   /**
-    * User's login name UI text field.
-    */
+   
     
     @FXML
     private Button btnSearch;
@@ -77,7 +82,10 @@ public class FTPController extends GenericController{
     private TreeTableColumn columnFile;
     
     private TreeItem<MyFile> rootItem;
-    
+    private Node rootIcon1 =  new ImageView(new Image(getClass().getResourceAsStream
+            ("/albergueperronclient/root.png")));
+    private Node rootIcon2 =  new ImageView(new Image(getClass().getResourceAsStream
+            ("/albergueperronclient/file.png"))); 
     
    
     /**
@@ -85,10 +93,11 @@ public class FTPController extends GenericController{
      * @param root The Parent object representing root node of view graph.
      */
     public void initStage(Parent root) throws IOException {
-        LOGGER.info("Initializing Login stage.");
+        
         //Create a scene associated to the node graph root.
         Scene scene = new Scene(root);
         //Associate scene to primaryStage(Window)
+        Stage stage = new Stage();
         stage.setScene(scene);
         //Set window properties
         stage.setTitle("FTP");
@@ -107,22 +116,27 @@ public class FTPController extends GenericController{
         treeFile.getSelectionModel().selectedItemProperty().addListener(this::itemSelected);
         
       
-        ftpManager.connect();
+        try {
+            ftpManager.connect();
+        } catch (FTPException ex) {
+            LOGGER.severe(ex.getMessage());
+            showErrorAlert("Error de conexión");
+        }
         
-        
-        rootItem = new TreeItem<MyFile>();
         MyFile rootF = new MyFile();
         rootF.setPath("/");
-        rootF.setFile(true);
+        rootF.setFile(false);
+        rootF.setName("root");
+        rootItem = new TreeItem<MyFile>(rootF,rootIcon1);
         rootItem.setExpanded(true);
 
-            // set the tree root
+        // set the tree root
         treeFile.setRoot(rootItem);
 
-            // start building the file tree
-        ftpManager.buildFileTree(treeFile.getRoot());
+        // start building the file tree
+        ftpManager.buildFileTree(rootItem);
         ftpManager.setTreeFile(treeFile);
-        //scene.add(treeFile);
+        
         stage.setScene(scene);
         //Show primary window
         
@@ -139,13 +153,24 @@ public class FTPController extends GenericController{
     private void handleWindowShowing(WindowEvent event){
         
            btnUpload.setDisable(true);
+           btnUpload.setMnemonicParsing(true);
+           btnUpload.setText("_Subir");
            btnDeleteF.setDisable(true);
+           btnDeleteF.setMnemonicParsing(true);
+           btnDeleteF.setText("_Borrar fichero");
            btnDownload.setDisable(true);
+           btnDownload.setMnemonicParsing(true);
+           btnDownload.setText("_Descargar");
            btnCrear.setDisable(true);
+           btnCrear.setMnemonicParsing(true);
+           btnCrear.setText("_Crear directorio");
            btnDeleteD.setDisable(true);
+           btnDeleteD.setMnemonicParsing(true);
+           btnDeleteD.setText("_Borrar directorio");
+           btnBack.setDisable(false);
+           btnBack.setMnemonicParsing(true);
+           btnBack.setText("_Volver");
           
-
-       
     }
     
     public void search(ActionEvent event){
@@ -159,33 +184,41 @@ public class FTPController extends GenericController{
         File selectedFile = fileChooser.showOpenDialog(stage);
         if (selectedFile != null) {
             txtFile.setText(selectedFile.getPath());
-            btnUpload.setDisable(false);
+            
                 
         }
        
     }
     
     public void upload(ActionEvent event){
-        
-        TreeItem<MyFile> itemSelected = (TreeItem<MyFile>) treeFile.getSelectionModel().getSelectedItem();
-        
-        
-        MyFile file=ftpManager.uploadFile(txtFile.getText());
-        TreeItem<MyFile> fileToUpload = new TreeItem<MyFile>(file);
-        itemSelected.getChildren().add(fileToUpload);
-        txtFile.setText("");
-        treeFile.refresh();
+        if(!txtFile.getText().equals("")){
+            TreeItem<MyFile> itemSelected = (TreeItem<MyFile>) treeFile.getSelectionModel().getSelectedItem();
+            MyFile file=ftpManager.uploadFile(txtFile.getText());
+            TreeItem<MyFile> fileToUpload = new TreeItem<MyFile>(file,rootIcon2);
+            itemSelected.getChildren().add(fileToUpload);
+            txtFile.setText("");
+            treeFile.refresh();
+        }else{
+           Alert alert = new Alert(Alert.AlertType.INFORMATION,"Elija un archivo local para subir");
+           alert.show();
+           
+        }
        
         
     }
       
       public void delete(ActionEvent event){
-       TreeItem<MyFile> treeItem= (TreeItem<MyFile>)treeFile.getSelectionModel().getSelectedItem();
-       TreeItem<MyFile> parent=treeItem.getParent();
-       //MyFile file =treeItem.getValue();
-       ftpManager.deleteFile(treeItem.getValue().getName());
-       parent.getChildren().remove(treeItem);
-        treeFile.refresh();
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION,"¿Quiere borrar el fichero?",
+                    ButtonType.CANCEL,ButtonType.OK);
+       
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.get() == ButtonType.OK){
+            TreeItem<MyFile> treeItem= (TreeItem<MyFile>)treeFile.getSelectionModel().getSelectedItem();
+            TreeItem<MyFile> parent=treeItem.getParent();
+            ftpManager.deleteFile(treeItem.getValue().getName());
+            parent.getChildren().remove(treeItem);
+            treeFile.refresh();
+        }
       }
             
     public void download(ActionEvent event){
@@ -200,7 +233,7 @@ public class FTPController extends GenericController{
         TreeItem<MyFile> itemSelected = (TreeItem<MyFile>) treeFile.getSelectionModel().getSelectedItem();
         
         MyFile file=ftpManager.createDirectory();
-        TreeItem<MyFile> directoryToCreate = new TreeItem<MyFile>(file);
+        TreeItem<MyFile> directoryToCreate = new TreeItem<MyFile>(file,rootIcon1);
         itemSelected.getChildren().add(directoryToCreate);
         treeFile.refresh();
         ftpManager.createDirectory();
@@ -209,12 +242,18 @@ public class FTPController extends GenericController{
     }
     
     public void deleteDir(ActionEvent event){
-       TreeItem<MyFile> treeItem= (TreeItem<MyFile>)treeFile.getSelectionModel().getSelectedItem();
-       TreeItem<MyFile> parent=treeItem.getParent();
-       ftpManager.deleteFile(treeItem.getValue().getName());
-       parent.getChildren().remove(treeItem);
-        treeFile.refresh();
-        ftpManager.deleteDirectory(treeItem.getValue().getPath());
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION,"¿Quiere borrar el directorio?",
+                    ButtonType.CANCEL,ButtonType.OK);
+       
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.get() == ButtonType.OK){
+            TreeItem<MyFile> treeItem= (TreeItem<MyFile>)treeFile.getSelectionModel().getSelectedItem();
+            TreeItem<MyFile> parent=treeItem.getParent();
+            ftpManager.deleteDirectory(treeItem.getValue().getPath());
+            parent.getChildren().remove(treeItem);
+            treeFile.refresh();
+            
+        }
          
         
     }
@@ -224,35 +263,32 @@ public class FTPController extends GenericController{
              Object oldValue,
              Object newValue) {
          
-            if(!newValue.equals(oldValue)){
-                 TreeItem<MyFile> selectedItem = (TreeItem<MyFile>) newValue;
+            if(newValue==null){
+                btnCrear.setDisable(true);
+                btnDeleteD.setDisable(true);
+                btnUpload.setDisable(true);
+                btnDownload.setDisable(true);
+                btnDeleteF.setDisable(true);  
+            }else{
+                TreeItem<MyFile> selectedItem = (TreeItem<MyFile>) newValue;
 
-                String path = null;
-                
-                if(selectedItem.getParent()!=null){
-                    path = selectedItem.getValue().getPath();
-                    ftpManager.changeDirectory(path);
-                     if(selectedItem.getValue().isFile()){
-                        btnCrear.setDisable(true);
-                        btnDeleteD.setDisable(true);
-                        btnUpload.setDisable(true);
-                        btnDownload.setDisable(false);
-                        btnDeleteF.setDisable(false);
-                    }else if(!selectedItem.getValue().isFile()){
-                        btnDownload.setDisable(true);
-                        btnDeleteF.setDisable(true);
-                        btnCrear.setDisable(false);
-                        btnDeleteD.setDisable(false);
-                        //if(txtFile.getText())
-                        btnUpload.setDisable(true);
-                    }
-
-                }else{
-                    path="/";
-                    ftpManager.changeDirectory(path);
+                String path = selectedItem.getValue().getPath();
+                ftpManager.changeDirectory(path);
+                if(selectedItem.getValue().isFile()){
+                    btnCrear.setDisable(true);
+                    btnDeleteD.setDisable(true);
+                    btnUpload.setDisable(true);
+                    btnDownload.setDisable(false);
+                    btnDeleteF.setDisable(false);
+                   
+                }else if(!selectedItem.getValue().isFile()){
+                    btnDownload.setDisable(true);
+                    btnDeleteF.setDisable(true);
+                    btnCrear.setDisable(false);
+                    btnDeleteD.setDisable(false);
+                    btnUpload.setDisable(false);
                 }
-               
-                
+      
             }
      }
